+++ kubernetes-client-api/src/test/java/io/fabric8/kubernetes/client/utils/HttpClientUtilsTest.java
@@ -189,6 +189,7 @@ class HttpClientUtilsTest {
           arguments("192.168.1.110", new String[] { "192.168.1.0/24" }),
           arguments("192.168.1.110", new String[] { "http://192.168.1.0/24" }),
           arguments("192.168.1.110", new String[] { "192.0.0.0/8" }),
+          arguments("2620:52:0:9c:0:0:0:1", new String[] { "2620:52:0:9c::/64" }),
           arguments("192.168.1.110", new String[] { "http://192.0.0.0/8" }));
     }