+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Fix #6747: Preventing websocket error logs when the client is closed
 * Fix #6781: Allowing ipv6 entries to work in NO_PROXY
+* Fix #6725: CRD generator missing type for GenericKubernetesResource
 
 #### Improvements